requirements.txt
streamlit
pandas
sentence-transformers
docx2txt
pdfminer.six
nltk
